RDKB-305: Issue with disabling/excluding CcspLogAgent from the firmware
Reason for change: When the ccsp-logagent is disabled or removed from a 
firmware build, it is observed that the bus-handles for RDKLogLevel and 
RDKLogEnable(and their component specific variables) have wrong values 
assigned to them(Device.LogAgent. does not exist) because of which the 
ccsp logs are not generated as expected despite being properly defined 
in the debug.ini file.

Test Procedure:
1. Disable the ccsp-logagent in the build.
2. Query for Device.LogAgent.
3. Confirm that the Destination Component cannot be found.
4. True to generate logs for any Ccsp Component using the ccsp-logagent 
by changing the log levels as required on debug.ini

Risks: None

Signed-Off by: Ajay Vishnu Bhat <abhat.contractor@libertyglobal.com>

Change-Id: I6bd95efab589b85555d6facf3b154fdd2a736e69