call HAL APIs for changes to Device.Ethernet.Interface.X.EEEEnable ( OFW-3733 )
Fix default value of Device.Ethernet.Interface.1.EEEEnable should be TRUE

Previously, any change to Device.Ethernet.Interface.X.EEEEnable
resulted in the new value being stored in PSM, but not being passed
to the HAL. Add appropriate calls to HAL and refactor code to store
values in PSM.

With the change, on reboot value is obtained from PSM and set in HAL
layer. On getv, value is read from PSM and on setv value is stored in
PSM and set in the HAL layer.