fix 2.5G ETH port Link detection failed when Modem with bridge mode ( OFW-4539 )
[Issue]
Wan port is down when the modem is in bridge mode

[Solution]
Set admin status of the port only if it is explicitly changed

[Testing Performed]
- Set mode to bridge mode and wait for the CPE to boot up
- confirm that wanoe port (eth3) is UP

Upstream Candidate: Maybe, but can not be upstreamed without full
details if how to test on a Comcast platform. Initial attempt to
upstream has been abandoned because description / testing details
were not sufficient.

Signed-off-by: Devin Mungan <dmungan@libertyglobal.com>