RDKBDEV-1860 : Dynamic population of Ethernet.Interface table
Reason for change: To replace the usage of hardcoded names and values with
runtime fetching of the actual names and values used by the platform from
a configuration file

Test Procedure: Update BBHM file with definitions of Ethernet interfaces
(dmsb.ethagent.if.%d). Ensure that dmcli shows Ethernet.Interface table
with interfaces defined in the configuration file.

Risks: low

Signed-off-by: Nikolay Papenkov <npapenkov@maxlinear.com>
Change-Id: I937f5cf9cf2e275946239fbfc4d1429814f5c06b